refactor(rendering): streamline error handling and remove unused components
- Removed LoadingSpinner and DefaultLoading components from component_eval_setup.js.
- Updated error handling in component_render.js to log errors without rendering HTML error messages.
- Simplified HTML rendering logic in html_render.js by returning empty strings on errors.
- Refactored component_render_setup.js to improve fallback content handling.
- Introduced HttpRuntimeClient in http-runtime-client.ts for better runtime client management.
- Removed NavigationErrorOverlay component and related logic from ClientRouter.tsx.
- Updated package.json to include sideEffects for better tree-shaking.
