+++ crates/rari/src/rsc/rendering/core/js/component_eval_setup.js
@@ -46,55 +46,6 @@ if (typeof globalThis.jsx === 'undefined')
 if (typeof globalThis.jsxs === 'undefined')
   globalThis.jsxs = createJsxDelegate(globalThis['~react']?.jsxRuntime?.jsxs, undefined)
 
-if (typeof globalThis.LoadingSpinner === 'undefined') {
-  if (
-    typeof document !== 'undefined'
-    && typeof document.getElementById === 'function'
-    && !document.getElementById('spinner-keyframes')
-  ) {
-    const head = document.head || document.getElementsByTagName('head')[0]
-    if (head) {
-      const style = document.createElement('style')
-      style.id = 'spinner-keyframes'
-      style.textContent = '@keyframes spin { from { transform: rotate(0deg); } to { transform: rotate(360deg); } }'
-      head.appendChild(style)
-    }
-  }
-
-  globalThis.LoadingSpinner = function () {
-    if (!globalThis.React || typeof globalThis.React.createElement !== 'function')
-      return null
-
-    return globalThis.React.createElement('div', {
-      style: {
-        width: '40px',
-        height: '40px',
-        border: '4px solid #f3f4f6',
-        borderTop: '4px solid #3b82f6',
-        borderRadius: '50%',
-        animation: 'spin 1s linear infinite',
-      },
-    })
-  }
-}
-
-if (typeof globalThis.DefaultLoading === 'undefined') {
-  globalThis.DefaultLoading = function () {
-    if (!globalThis.React || typeof globalThis.React.createElement !== 'function' || !globalThis.LoadingSpinner)
-      return null
-
-    return globalThis.React.createElement('div', {
-      style: {
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        padding: '2rem',
-        minHeight: '200px',
-      },
-    }, globalThis.React.createElement(globalThis.LoadingSpinner))
-  }
-}
-
 if (!globalThis.readFileSync && globalThis['~node']?.modules && globalThis['~node'].modules.get) {
   const nodeFs = globalThis['~node'].modules.get('node:fs')
   if (nodeFs && nodeFs.readFileSync) {
