+++ crates/rari/src/rsc/rendering/streaming/js/component_render_setup.js
@@ -119,26 +119,13 @@ globalThis['~render'].componentAsync = async function () {
           }
         }
 
-        let fallbackContent
-        if (loadingComponent) {
-          if (
-            typeof loadingComponent === 'object'
-            && (loadingComponent.type || loadingComponent.$$typeof)
-          ) {
-            fallbackContent = loadingComponent
-          }
-          else {
-            fallbackContent = globalThis['~react'].originalCreateElement('div', {
-              className: 'rari-loading',
-              children: 'Loading...',
-            })
-          }
-        }
-        else {
-          fallbackContent = globalThis['~react'].originalCreateElement('div', {
-            className: 'rari-loading',
-            children: 'Loading...',
-          })
+        let fallbackContent = null
+        if (
+          loadingComponent
+          && typeof loadingComponent === 'object'
+          && (loadingComponent.type || loadingComponent.$$typeof)
+        ) {
+          fallbackContent = loadingComponent
         }
 
         registerBoundary(boundaryId, fallbackContent, null)
@@ -230,20 +217,13 @@ globalThis['~render'].componentAsync = async function () {
           }
         }
 
-        let fallbackContent
-        if (
+        const fallbackContent = (
           loadingComponent
           && typeof loadingComponent === 'object'
           && (loadingComponent.type || loadingComponent.$$typeof)
-        ) {
-          fallbackContent = loadingComponent
-        }
-        else {
-          fallbackContent = globalThis['~react'].originalCreateElement('div', {
-            className: 'rari-loading',
-            children: 'Loading...',
-          })
-        }
+        )
+          ? loadingComponent
+          : null
 
         if (!parentBoundaryId)
           registerBoundary(boundaryId, fallbackContent, null)
