+++ packages/rari/src/client.ts
@@ -8,9 +8,6 @@ export { NavigationErrorHandler } from './router/navigation-error-handler'
 
 export type { NavigationError, NavigationErrorHandlerOptions, NavigationErrorType } from './router/navigation-error-handler'
 
-export { NavigationErrorOverlay } from './router/NavigationErrorOverlay'
-export type { NavigationErrorOverlayProps } from './router/NavigationErrorOverlay'
-
 export {
   clearPropsCache,
   clearPropsCacheForComponent,
@@ -46,20 +43,6 @@ export type {
   RouteSegmentType,
   TemplateEntry,
 } from './router/types'
-
-export {
-  createErrorBoundary,
-  createHttpRuntimeClient,
-  createLoadingBoundary,
-  DefaultError,
-  DefaultLoading,
-  ErrorBoundary,
-  HttpRuntimeClient,
-  LoadingSpinner,
-  NotFound,
-} from './runtime-client'
-
-export type { RuntimeClient } from './runtime-client'
 
 export type {
   ServerConfig,
